requirements.txt
bcrypt==3.1.6
blinker==1.4
cffi==1.13.1
cov-core==1.15.0
coverage==3.7.1
cssmin==0.2.0
flake8==2.4.0
Flask==0.12.2
Flask-Assets==0.12
Flask-Bcrypt==0.7.1
Flask-DebugToolbar==0.11.0
Flask-DotEnv==0.1.2
Flask-Login==0.2.11
Flask-Mail==0.9.1
Flask-Migrate==2.0.1
Flask-Script==2.0.5
Flask-SQLAlchemy==2.1
Flask-WTF==0.11
itsdangerous==0.24
Jinja2==2.10
jsmin==2.1.1
MarkupSafe==1.1.0
mccabe==0.3
pep8==1.5.7
Pillow==6.2.1
psycopg2==2.8.4
py==1.7.0
pycparser==2.19
pyflakes==0.8.1
pytest==2.7.0
pytest-cov==1.8.1
python-dotenv==0.13.0
six==1.12.0
SQLAlchemy==1.2.17
webassets==0.12.1
Werkzeug==0.16.1
WTForms==2.2.1
